requirements.txt
flask
requests
google-generativeai
twilio
gunicorn